AB Testing
In this notebook, we will show how to conduct A/B testings using a two sample Z test. There is a drug A in the market that removes the headache in X number of hours. New drug B is being developed in this same pharma company and their hypothesis is that B can reduce the headache in less number of hours (i.e Recovery time for drug B is better compared to drug A). We have conducted the clinical trail and the results are available in the csv file.